Add custom rule for dropping the dynamic collection on out method

+++ lib/mongoid/report/output.rb
@@ -2,15 +2,23 @@ module Mongoid
   module Report
 
     class Output
-      attr_accessor :collection_name
+      attr_accessor :collection_name, :options
 
       def do(rows)
-        collection.drop()
+        drop()
         collection.insert(rows)
       end
 
       def present?
         collection_name.present?
+      end
+
+      def drop
+        return collection.drop() unless options[:drop].present?
+
+        # We will use custom way for dropping the collection or removing the
+        # records partially
+        collection.find(options[:drop]).remove()
       end
 
       private
